Fixing (see description)
- Fixed deltaTime
- Fixed the bug "crash sometimes when the game is closed". The destructor of Window was not called because not declared virtual
- Create CoreMinimal.h, include all the essential Engine headers
- Created as well Enums.h and ForwardDeclarations.h

+++ Engine/Camera3D.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include <glm.hpp>
+#include "CoreMinimal.h"
 
-#include "GLSLProgram.h"
 
 namespace Engine {
 
@@ -28,7 +27,7 @@ namespace Engine {
             m_NeedUpdate = true;
         }
 
-        void setScale(int scale) {
+        void setScale(float scale) {
             m_Scale = scale;
             m_NeedUpdate = true;
         }
